Add refis configuration file
