+++ src/scheduler/server.ts
@@ -1,25 +1,11 @@
 import { Job, Queue, QueueScheduler, Worker } from 'bullmq';
-import Redis from 'ioredis';
-import * as Monitoring from '../monitoring';
 
+import * as Monitoring from '../monitoring';
 import logger from '../logger';
 import refreshOrganizations from '../jobs/refreshOrganizations';
+import options from '../config/redis';
 
 const SCHEDULER_QUEUE = 'ServerSchedulerQueue';
-
-const redisOptions: Redis.RedisOptions = {
-  maxRetriesPerRequest: null,
-  enableReadyCheck: false,
-};
-
-const connection = new Redis(
-  process.env.REDIS_URL || 'redis://127.0.0.1:6379',
-  redisOptions,
-);
-
-const options = {
-  connection,
-};
 
 const scheduler = new QueueScheduler(SCHEDULER_QUEUE, options);
 const schedulerQueue = new Queue(SCHEDULER_QUEUE, options);
